requirements.txt
flask==2.1.2
mysql.connector==2.2.9